WIP: support select box in questions
Supports a select box as an answer in questions, to handle cases like business type with many possible answers.

Test plan:

Added an input like this to an existing question:

    select: {
      name: "q1_select_opt",
      options: {
        a: "Option A",
        b: "Option B"
      }
    }

I selected on of the options, and saw the next question was now answerable.  Verified the answer was correctly
recorded in the JS console:

    answers["q1_is_non_profit_reg_in_nj"]
    "a"